// frontend/lib/api/generated/getForManagementStatus.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */

export type GetForManagementStatus =
  (typeof GetForManagementStatus)[keyof typeof GetForManagementStatus];

export const GetForManagementStatus = {
  SUBMITTED: "SUBMITTED",
  IN_PROGRESS: "IN_PROGRESS",
  REJECTED: "REJECTED",
  FEEDBACKED: "FEEDBACKED",
} as const;